move api to inner dir and update test
