+++ .github/workflows/tests.yml
@@ -30,7 +30,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install pytest httpx
+          pip install pytest httpx pytest-repeat
           pip install -r requirements.txt
       
       - name: install apt dependecies
